chore: decrease test data, and rename utils to testData
